BBlustering

BBlustering

MCBB.bbcluster — Function.

bbcluster(D::AbstractDistanceMatrix, dplus::AbstractVector{T}, dminus::AbstractVector{T}, pars::AbstractVector{T}, p_eps::T, minpts::Int; k::S=1.5, par_distance_func::Union{Function, Nothing}=nothing) where {T,S}<:Real

Performs the BBClustering, a modified DBSCAN clustering adjusted for Basin Bifurcation Analysis

Inputs:

D: Distance Matrix (NxN)

dplus: Response of Distance Measure at p+\delta p

dminus: Response of Distance Measure at p-\delta p

pars: Parameter vector

delta_p: Used to estimate response dplus and dminus

p_eps: Epsilon Parameter, only points with parameters closer than p_eps are connected.

minpts: Minimum number of points for a cluster, otherwise outlier

k: Paramater for the clustering, should be 1 < k < 2

par_distance_func: Distance function for parameters, check: par_distance_func(pars[i],pars[j]) < p_eps

bbcluster(D::AbstractArray, prob::MCBBProblem, sol::MCBBSol, deltap::T; peps::Union{Nothing,T}=nothing, minpts::Int=1, k::Number=1.5, pardistancefunc::Union{Function,Nothing}=nothing) where T<:Real

Convenience wrapper of the above defined function with 'MCBBProblem' and 'MCBBSol' as inputs. Default value for p_eps is five times the mean parameter difference.

source

MCBB.BBClusterResult — Type.

BBClusterResult <: ClusteringResult

Contains the solutions of the bbcluster. Has the fields seeds, assignments and counts.

source