Pre-Made Systems and Parameters

There will be a list of all pre-made systems and parameters here.

MCBB.DEParameters — Type

DEParameters

Abstract type for all parameter types. If you define your own systems the parameters have to have this as a supertype.

MCBB.logistic — Function

logistic(u_next, u, p::logistic_parameters, t)

Logistic Map: $x_{n+1} = rx_n (1 - x_n)$

MCBB.logistic_parameters — Type

 logistic_parameters

Parameters of the logistic map. Its function is logistic.

Fields: *r::Float64

MCBB.henon — Function

henon(u_next, u, p::henon_parameters, t)

Henon map. $x_{n+1} = 1 - a x_{n}^2 + y_{n}$ $y_{n+1} = b x_{n}$

MCBB.henon_parameters — Type

henon_parameters

Parameters of the Henon map. Its function is henon.

Fields: *a::Float64 *b::Float64

MCBB.kuramoto — Function

kuramoto(du, u, p::kuramoto_parameters, t)

First order Kuramoto system with all-to-all coupling. $\dot{\theta}_i = w_i + K/N \sum_{i} \sin(\theta_j - \theta_i)$

MCBB.kuramoto_parameters — Type

kuramoto_parameters

Parameters of a first order Kuramoto system with all-to-all coupling. Fields:

K::Number: Coupling Strength

w::Array{Float64}: Eigenfrequencies

N::Int: Number of Oscillators

MCBB.kuramoto_network — Function

kuramoto_network(du, u, p::kuramoto_parameters, t)

First order Kuramoto system on a network. $\dot{\theta}_i = w_i + K/N \sum_{i} A_{ij}\sin(\theta_j - \theta_i)$

MCBB.kuramoto_network_parameters — Type

kuramoto_network_parameters

Parameters of a first order Kuramoto system on a network. Fields:

K::Number: Coupling Strength

w::Array{Float64}: Eigenfrequencies

N::Int: Number of Oscillators

A: Adjacency matrix

MCBB.second_order_kuramoto — Function

second_order_kuramoto(du, u, p::second_order_kuramoto_parameters, t)

Second order Kuramoto system on the adjacency matrix $A_{ij} = E'_{ie} E_{ej}$.

$\dot{\theta}_i = w_i$ $\dot{\omega} = \Omega_i - \alpha\omega + \lambda\sum_{j=1}^N A_{ij} sin(\theta_j - \theta_i)$

MCBB.second_order_kuramoto_parameters — Type

second_order_kuramoto_parameters

Fields:

systemsize::Int, number of oscillators

damping::Float64, Damping, also referred to as $\alpha$

coupling::Float64, Coupling strength, also referred to as $\lambda$

incidence, oriented incidence matix of the network, also referred to as $\E_{ei}$

indicating if a vertex $i$ belongs to an edge $e$. Following the conventions of LightGraphs

drive::Array{Float64}, external driving, also referred to as $\Omega$

MCBB.second_order_kuramoto_chain — Function

second_order_kuramoto_chain(du, u, p::kuramoto_parameters, t)

Second order Kuramoto system on a chain.

$\dot{\theta}_i = w_i$ $\dot{\omega} = \Omega_i - \alpha\omega + \lambda\sum_{j=1}^N A_{ij} sin(\theta_j - \theta_i)$

$A_{ij} = 1$ if and only if $|i-j|=1$, otherwise $A_{ij}=0$

MCBB.second_order_kuramoto_chain_parameters — Type

second_order_kuramoto_chain_parameters

Fields:

systemsize::Int, number of oscillators

damping::Float64, Damping, also referred to as $\alpha$

coupling::Float64, Coupling strength, also referred to as $\lambda$

drive::Array{Float64}, external driving, also referred to as $\Omega$

MCBB.non_local_kuramoto_ring — Function

non_local_kuramoto_ring(du, u, p::non_local_kuramoto_ring_parameters, t)

First order Kuramoto oscillators on a ring with non-local coupling.

$\frac{\theta_k}{dt} = \omega_0 - \sum_{j=1}^N G_1\left(\frac{2\pi}{N}(k-j)\right)\sin\left(\theta_k(t) - \theta_j(k) + \alpha\right)$

MCBB.non_local_kuramoto_ring_parameters — Type

non_local_kuramoto_ring_parameters <: DEParameters

N::Integer: Number of Oscillators

fak::Float64: $\frac{2\pi}{n}$

omega_0::Number: eigenfrequency of all oscillators

phase_delay: Phase Delay $\alpha$

coupling_function: Coupling function $G(x)$, signature (x::Number)-> value::Number

MCBB.order_parameter — Function

order_parameter(u::AbstractArray, N::Int)

Order Parameter of a Kuramoto System

MCBB.roessler_parameters — Type

roessler_parameters <: DEParameters

Parameters of a Roessler network

a::Array{Float64}: a Parameters of all oscillators.

b::Array{Float64}: b Parameters of all oscillators.

c::Array{Float64}: c Parameters of all oscillators.

K::Float64: Coupling Strength

L::Array{Float64}: Laplacian matrix of the network

N::Int: Number of nodes/oscilattors

roessler_parameters(a, b, c, K, k::Int64, p::Float64, N)

Generates a set of roessler_parameters with a Watts Strogatz random Network with mean degreee k and rewiring probability p

MCBB.roessler_network — Function

roessler_network(du, u, p::roessler_parameters, t)

N Roessler Parameters coupled on their x-component